Ændret typen på Compresor.m_rawSize og m_compressedSize fra size_t til UINT64

+++ Test/Compress/Compressor.h
@@ -7,7 +7,7 @@ class Compressor {
 protected:
   static const int endOfStream;
 
-  size_t m_compressedSize, m_rawSize;
+  UINT64 m_compressedSize, m_rawSize;
 public:
   virtual void compress(ResetableByteInputStream &input, ByteOutputStream &output, int quality = 3) = 0;
   virtual void expand(  ByteInputStream          &input, ByteOutputStream &output) = 0;
